+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 Change Log
 ==========
 
-## 0.19.1 **UNRELEASED**
+## 0.20.1 **UNRELEASED**
+
+
+## 0.20.0 *(2023-11-17)*
 
 ### Navigation
 
@@ -11,6 +14,7 @@ Change Log
 can be overriden with `NavHostDefaults.transitionAnimations` or disabled with
 `NavHostTransitionAnimations.noAnimations`. Default animations are the same as default animations
 in AndroidX's `NavHost`.
+
 
 ## 0.19.0 *(2023-11-09)*